refactor : 상태관리 적용
