+++ src/store/store.ts
@@ -25,3 +25,29 @@ export const useApiUrlStore = create<ApiUrlState>((set) => ({
   // apiUrl: 'http://localhost:8080/api',
   setApiUrl: (url: string) => set((state) => ({ ...state, apiUrl: url })),
 }))
+
+interface UserListState {
+  userList: UserList[]
+  setUserList: (userList: UserList[]) => void
+}
+
+interface UserList {
+  id: number
+  name: string
+  nickname: string
+  part: string
+  email: string
+  interests: string
+  blogUrl: string
+  publicRelations: string
+  job: string
+  heart: number
+  starAverage: number
+  solved: number
+  matchingCount: number
+}
+
+export const useUserListStore = create<UserListState>((set) => ({
+  userList: [],
+  setUserList: (userList) => set({ userList }),
+}))
